Refactoring: Integrations (#318)
* 2.0.4-alpha.0

* fix(caps): refetch caps on mount or arg change. (#307)

* fix: unified limit exhausted message across app & adjusted text

* fix: better wording

* 2.0.4-alpha.1

* fix: updated offline message

* fix: increased size for input field

* fix: if only one chat tab left and active tab is dashboard, allowing click on chat tab

* feat: upgrade to pro button in main dropdown & reworked links in dropdown

* fix: show plugin settings if jetbrains, otherwise extension settings

* chore: revert deletion of integrations

* chore: revert deletion of integrations

* fix: startup keyword shouldn't be null

* wip(updated tour): integrations step & updated refs

* feat: updated onboarding with new data & small adjustments to the UI

* 2.0.4-alpha.2

* wip(checkpoints): ui for popup and user input

* wip(checkpoints): truncation of links for files

* wip(checkpoints): storing checkpoints for user messages

* wip(checkpoints): API handler, typeguards, saving restore results to store and fixed fixtures

* wip(checkpoints): fixed types issue

* wip(checkpoints): fix of tests

* wip(checkpoints): stubs for checkpoints & storybook

* feat(checkpoints): undo functionality & refactoring

* chore: fixed types for CI

* refactoring(checkpoints): extracted logic and functions to useCheckpoints hook

* chore: parsing error text adjusted

* fix: make checkpoints optional & hiding revert changes button if there are no checkpoints

* fix: support esc key press & click out of the dialog to close it

* chore: removed convertUTCToLocalTime() util

* fix: remove hardcoded locale en-GB to use system one

* feat: reworked logic & improved UI/UX

* chore: better wording with dates

* fix: checking for index of of current message instead of checking for user messages amount

* fix(onboarding): updated GIF and links for final screen

* fix: updated tour ending

* 2.0.4-alpha.3

* fix: fixed validation of paths for puzzle links on Windows, adjusted texts

* improve(puzzle_links): extractLinkFromPuzzle unit tests coverage

* refactoring(integrations): useIntegrations hook to isolate all logic out of the component

* refactoring(integrations): simplify internal code for IntegrationForm, IntegrationView, simplified removed renderIntegrationField and using component now, created helper hooks to minimize file sizes and increase readability

* refactoring(integrations): removed renderIntegrationField function, IntegrationFormField component created instead

* refactoring(integrations): better renderSmartlinks()

* refactoring(integrations): removed inline call of functions, created separate components instead

* refactoring(integrations): better folder structure
