+++ src/components/IntegrationsView/IntegrationForm/IntegrationForm.tsx
@@ -1,33 +1,26 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
+import { Button, Flex, Grid, Text } from "@radix-ui/themes";
 import classNames from "classnames";
-import { useGetIntegrationDataByPathQuery } from "../../../hooks/useGetIntegrationDataByPathQuery";
-
-import type { FC, FormEvent, Dispatch } from "react";
-import type {
-  Integration,
-  IntegrationField,
-  IntegrationPrimitive,
-  ToolConfirmation,
-} from "../../../services/refact";
-
-import styles from "./IntegrationForm.module.css";
-import { Spinner } from "../../Spinner";
-import { Badge, Button, Flex, Grid, Heading, Text } from "@radix-ui/themes";
-import { IntegrationDocker } from "../IntegrationDocker";
-import { SmartLink } from "../../SmartLink";
-import { renderIntegrationFormField } from "../../../features/Integrations/renderIntegrationFormField";
-import { IntegrationAvailability } from "./IntegrationAvailability";
-import { IntegrationDeletePopover } from "../IntegrationDeletePopover";
+import { FC, FormEvent, useEffect } from "react";
 import { debugIntegrations } from "../../../debugConfig";
+import { useGetIntegrationDataByPathQuery } from "../../../hooks/useGetIntegrationDataByPathQuery";
+import { Spinner } from "../../Spinner";
+import { Confirmation } from "../Confirmation";
+import { useFormAvailability } from "../hooks/useFormAvailability";
+import { useFormFields } from "../hooks/useFormFields";
+import { IntegrationDocker } from "../IntegrationDocker";
+
 import type { ToolParameterEntity } from "../../../services/refact";
 import {
   areAllFieldsBoolean,
   areToolConfirmation,
-  areToolParameters,
+  type Integration,
+  type ToolConfirmation,
 } from "../../../services/refact";
-import { Confirmation } from "../Confirmation";
-import { ExclamationTriangleIcon } from "@radix-ui/react-icons";
-import { useEventsBusForIDE } from "../../../hooks";
+import { ErrorState } from "./ErrorState";
+import { FormAvailabilityAndDelete } from "./FormAvailabilityAndDelete";
+import { FormFields } from "./FormFields";
+import { FormSmartlinks } from "./FormSmartlinks";
+import styles from "./IntegrationForm.module.css";
 
 type IntegrationFormProps = {
   integrationPath: string;
@@ -41,11 +34,11 @@ type IntegrationFormProps = {
   handleChange: (event: FormEvent<HTMLFormElement>) => void;
   onSchema: (schema: Integration["integr_schema"]) => void;
   onValues: (values: Integration["integr_values"]) => void;
-  setAvailabilityValues: Dispatch<
+  setAvailabilityValues: React.Dispatch<
     React.SetStateAction<Record<string, boolean>>
   >;
-  setConfirmationRules: Dispatch<React.SetStateAction<ToolConfirmation>>;
-  setToolParameters: Dispatch<
+  setConfirmationRules: React.Dispatch<React.SetStateAction<ToolConfirmation>>;
+  setToolParameters: React.Dispatch<
     React.SetStateAction<ToolParameterEntity[] | null>
   >;
   handleSwitchIntegration: (
@@ -71,146 +64,65 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
   setToolParameters,
   handleSwitchIntegration,
 }) => {
-  const [areExtraFieldsRevealed, setAreExtraFieldsRevealed] = useState(false);
-
   const { integration } = useGetIntegrationDataByPathQuery(integrationPath);
-  const { openFile } = useEventsBusForIDE();
 
-  const handleAvailabilityChange = useCallback(
-    (fieldName: string, value: boolean) => {
-      setAvailabilityValues((prev) => ({ ...prev, [fieldName]: value }));
-    },
-    [setAvailabilityValues],
-  );
+  const {
+    importantFields,
+    extraFields,
+    areExtraFieldsRevealed,
+    toggleExtraFields,
+  } = useFormFields(integration.data?.integr_schema.fields);
+  const {
+    handleAvailabilityChange,
+    handleConfirmationChange,
+    handleToolParameters,
+  } = useFormAvailability({
+    setAvailabilityValues,
+    setConfirmationRules,
+    setToolParameters,
+  });
 
-  const handleConfirmationChange = useCallback(
-    (fieldName: string, values: string[]) => {
-      setConfirmationRules((prev) => {
-        return { ...prev, [fieldName as keyof ToolConfirmation]: values };
-      });
-    },
-    [setConfirmationRules],
-  );
-
-  const handleToolParameters = useCallback(
-    (value: ToolParameterEntity[]) => {
-      setToolParameters(value);
-    },
-    [setToolParameters],
-  );
-
-  const shouldIntegrationFormBeDisabled = useMemo(() => {
-    if (!integration.data?.integr_values) return false;
-    return isDisabled;
-  }, [isDisabled, integration]);
-
-  useEffect(() => {
-    if (
-      integration.data?.integr_values &&
-      integration.data.integr_values.available &&
-      typeof integration.data.integr_values.available === "object" &&
-      areAllFieldsBoolean(integration.data.integr_values.available)
-    ) {
-      Object.entries(integration.data.integr_values.available).forEach(
-        ([key, value]) => {
-          handleAvailabilityChange(key, value);
-        },
-      );
-    }
-  }, [integration, handleAvailabilityChange]);
-
-  useEffect(() => {
-    if (
-      integration.data?.integr_values &&
-      integration.data.integr_values.parameters &&
-      Array.isArray(integration.data.integr_values.parameters) &&
-      areToolParameters(integration.data.integr_values.parameters)
-    ) {
-      handleToolParameters(integration.data.integr_values.parameters);
-    }
-  }, [integration, handleToolParameters]);
-
+  // Set initial values from integration data
   useEffect(() => {
     if (integration.data?.integr_schema) {
       onSchema(integration.data.integr_schema);
     }
-
     if (integration.data?.integr_values) {
       onValues(integration.data.integr_values);
     }
-    debugIntegrations(`[DEBUG]: integration.data: `, integration);
-  }, [integration, onSchema, onValues]);
 
-  const importantFields = Object.entries(
-    integration.data?.integr_schema.fields ?? {},
-  )
-    .filter(([_, field]) => !field.f_extra)
-    .reduce<
-      Record<string, IntegrationField<NonNullable<IntegrationPrimitive>>>
-    >((acc, [key, field]) => {
-      acc[key] = field;
-      return acc;
-    }, {});
+    debugIntegrations("[DEBUG]: integration.data: ", integration);
+  }, [integration, onSchema, onValues, handleAvailabilityChange]);
 
-  const extraFields = Object.entries(
-    integration.data?.integr_schema.fields ?? {},
-  )
-    .filter(([_, field]) => field.f_extra)
-    .reduce<
-      Record<string, IntegrationField<NonNullable<IntegrationPrimitive>>>
-    >((acc, [key, field]) => {
-      acc[key] = field;
-      return acc;
-    }, {});
+  useEffect(() => {
+    if (
+      integration.data?.integr_values?.available &&
+      areAllFieldsBoolean(integration.data.integr_values.available)
+    ) {
+      Object.entries(integration.data.integr_values.available).forEach(
+        ([fieldKey, value]) => {
+          handleAvailabilityChange(fieldKey, value);
+        },
+      );
+    }
+  }, [integration, handleAvailabilityChange]);
 
   if (integration.isLoading) {
     return <Spinner spinning />;
   }
 
   if (!integration.data) {
-    return (
-      <div>
-        <p>No integration found</p>
-      </div>
-    );
+    return <Text>No integration found</Text>;
   }
 
   if (integration.data.error_log.length > 0) {
-    const errorMessage = integration.data.error_log[0].error_msg;
-    const integrationFile = integration.data.error_log[0].integr_config_path;
-    const errorLine = integration.data.error_log[0].error_line;
     return (
-      <Flex width="100%" direction="column" align="start" gap="4">
-        <Text size="2" color="gray">
-          Whoops, this integration has a syntax error in the config file. You
-          can fix this problem by editing the config file.
-        </Text>
-        <Badge size="2" color="red">
-          <ExclamationTriangleIcon />
-          {errorMessage}
-        </Badge>
-        <Flex align="center" gap="2">
-          <Button
-            variant="outline"
-            color="gray"
-            onClick={() =>
-              openFile({
-                file_name: integrationFile,
-                line: errorLine === 0 ? 1 : errorLine,
-              })
-            }
-          >
-            Open {integration.data.integr_name}.yaml
-          </Button>
-          <IntegrationDeletePopover
-            integrationName={integration.data.integr_name}
-            integrationConfigPath={integration.data.integr_config_path}
-            isApplying={isApplying}
-            isDeletingIntegration={isDeletingIntegration}
-            handleDeleteIntegration={handleDeleteIntegration}
-          />
-        </Flex>
-      </Flex>
+      <ErrorState
+        integration={integration.data}
+        onDelete={handleDeleteIntegration}
+        isApplying={isApplying}
+        isDeletingIntegration={isDeletingIntegration}
+      />
     );
   }
 
@@ -221,6 +133,7 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
           {integration.data.integr_schema.description}
         </Text>
       )}
+
       <form
         onSubmit={handleSubmit}
         onChange={handleChange}
@@ -228,114 +141,37 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
       >
         <Flex direction="column" gap="2">
           <Grid mb="0">
-            {integration.data.integr_values && (
-              <Flex align="start" justify="between">
-                <Flex
-                  gap="4"
-                  mb="4"
-                  align="center"
-                  justify="between"
-                  className={styles.switchInline}
-                >
-                  {integration.data.integr_values.available &&
-                    Object.keys(integration.data.integr_values.available).map(
-                      (key) => (
-                        <IntegrationAvailability
-                          key={key}
-                          fieldName={key}
-                          value={availabilityValues[key]}
-                          onChange={handleAvailabilityChange}
-                        />
-                      ),
-                    )}
-                </Flex>
-                <IntegrationDeletePopover
-                  integrationName={integration.data.integr_name}
-                  integrationConfigPath={integration.data.integr_config_path}
-                  isApplying={isApplying}
-                  isDeletingIntegration={isDeletingIntegration}
-                  handleDeleteIntegration={handleDeleteIntegration}
-                />
-              </Flex>
-            )}
-            {integration.data.integr_schema.smartlinks &&
-              integration.data.integr_schema.smartlinks.length > 0 && (
-                <Flex width="100%" direction="column" gap="1" mb="6">
-                  <Flex align="center" gap="3" mt="2" wrap="wrap">
-                    <Heading as="h6" size="2" weight="medium">
-                      Actions:
-                    </Heading>
-                    {integration.data.integr_schema.smartlinks.map(
-                      (smartlink, index) => {
-                        return (
-                          <SmartLink
-                            key={`smartlink-${index}`}
-                            smartlink={smartlink}
-                            integrationName={
-                              integration.data?.integr_name ?? ""
-                            }
-                            integrationProject={
-                              integration.data?.project_path ?? ""
-                            }
-                            integrationPath={
-                              integration.data?.integr_config_path ?? ""
-                            }
-                            shouldBeDisabled={
-                              smartlink.sl_enable_only_with_tool
-                                ? integration.data?.integr_values === null ||
-                                  !shouldIntegrationFormBeDisabled ||
-                                  !availabilityValues.on_your_laptop
-                                : false
-                            }
-                          />
-                        );
-                      },
-                    )}
-                  </Flex>
-                </Flex>
-              )}
-            <Grid gap="2" className={styles.gridContainer}>
-              {Object.keys(importantFields).map((fieldKey) => {
-                if (integration.data) {
-                  return renderIntegrationFormField({
-                    fieldKey: fieldKey,
-                    values: integration.data.integr_values,
-                    field: integration.data.integr_schema.fields[fieldKey],
-                    integrationName: integration.data.integr_name,
-                    integrationPath: integration.data.integr_config_path,
-                    integrationProject: integration.data.project_path,
-                    onToolParameters: handleToolParameters,
-                  });
-                }
-              })}
-              {Object.keys(extraFields).map((fieldKey) => {
-                if (integration.data) {
-                  return renderIntegrationFormField({
-                    fieldKey: fieldKey,
-                    values: integration.data.integr_values,
-                    field: integration.data.integr_schema.fields[fieldKey],
-                    integrationName: integration.data.integr_name,
-                    integrationPath: integration.data.integr_config_path,
-                    integrationProject: integration.data.project_path,
-                    isFieldVisible: areExtraFieldsRevealed,
-                    onToolParameters: handleToolParameters,
-                  });
-                }
-              })}
-            </Grid>
+            <FormAvailabilityAndDelete
+              integration={integration.data}
+              availabilityValues={availabilityValues}
+              handleAvailabilityChange={handleAvailabilityChange}
+              isApplying={isApplying}
+              isDeletingIntegration={isDeletingIntegration}
+              onDelete={handleDeleteIntegration}
+            />
+            <FormSmartlinks
+              integration={integration.data}
+              smartlinks={integration.data.integr_schema.smartlinks}
+              availabilityValues={availabilityValues}
+            />
+            <FormFields
+              integration={integration.data}
+              importantFields={importantFields}
+              extraFields={extraFields}
+              areExtraFieldsRevealed={areExtraFieldsRevealed}
+              onToolParameters={handleToolParameters}
+            />
           </Grid>
-          {Object.values(extraFields).length > 0 && (
+
+          {Object.keys(extraFields).length > 0 && (
             <Button
               variant="soft"
               type="button"
               color="gray"
               size="2"
-              onClick={() => setAreExtraFieldsRevealed((prev) => !prev)}
+              onClick={toggleExtraFields}
               mb="1"
-              mt={{
-                initial: "3",
-                xs: "0",
-              }}
+              mt={{ initial: "3", xs: "0" }}
               className={styles.advancedButton}
             >
               {areExtraFieldsRevealed
@@ -343,14 +179,14 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
                 : "Show advanced configuration"}
             </Button>
           )}
+
           {!integration.data.integr_schema.confirmation.not_applicable && (
             <Flex gap="4" mb="3">
               <Confirmation
                 confirmationByUser={confirmationRules}
                 confirmationFromValues={
-                  integration.data.integr_values !== null &&
                   areToolConfirmation(
-                    integration.data.integr_values.confirmation,
+                    integration.data.integr_values?.confirmation,
                   )
                     ? integration.data.integr_values.confirmation
                     : null
@@ -362,6 +198,7 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
               />
             </Flex>
           )}
+
           <Flex
             justify="end"
             width="100%"
@@ -375,20 +212,11 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
                 variant="solid"
                 type="submit"
                 size="2"
-                title={
-                  shouldIntegrationFormBeDisabled
-                    ? "Cannot apply, no changes made"
-                    : "Apply"
-                }
-                className={classNames(
-                  {
-                    [styles.disabledButton]:
-                      isApplying || shouldIntegrationFormBeDisabled,
-                  },
-                  styles.button,
-                  styles.applyButton,
-                )}
-                disabled={shouldIntegrationFormBeDisabled}
+                title={isDisabled ? "Cannot apply, no changes made" : "Apply"}
+                className={classNames(styles.button, styles.applyButton, {
+                  [styles.disabledButton]: isApplying || isDisabled,
+                })}
+                disabled={isDisabled}
               >
                 {isApplying ? "Applying..." : "Apply"}
               </Button>
@@ -396,6 +224,7 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
           </Flex>
         </Flex>
       </form>
+
       {integration.data.integr_schema.docker && (
         <Flex mt="6" direction="column" align="start" gap="5">
           <IntegrationDocker
